Reorganize file structure. Make GeoSource, PolygonLayers, PointLayers folders
