+++ app/src/app/components/Map/PointLayers/PointSelectionLayer.tsx
