+++ app/src/app/components/Map/PolygonLayers/GeometryBackgroundLayer.tsx
