+++ app/src/app/components/Map/PolygonLayers/GeometryOutlineLayer.tsx
